zmq_adapter/requirements.txt
pyzmq
formant